// WebAPIAngular.Angular/src/app/model/aPDIRCircularAfter.ts
export class APDIRCircularAfter {
    APDIRCircularAfterId: number;
    APDIRCircularParentId: number;
    APDIRCircularId: number;
    APDIRCircularNo: string;
    APDIRCircularName: string;
    APDIRCircularDate: string;
    APDIRCircularEffectiveDate: string;
    Year: string;
    APDIRCircularYearName: string;
    PDF: string;
    IsDeleted: boolean;
    IsActive: boolean;
    CreatedDate: string;
    ModifiedDate: string;
}

export class GetAPDIRCircularAfterRequest {
    APDIRCircularAfterId: number;
    APDIRCircularParentId: number;
    SearchText: string;
    IsActive: boolean;
    PageNumber: number;
    PageSize: number;
    OrderBy: string;
    OrderByDirection: string;
}